READ.ME
The biggest challenges I have is being home and not able to go to class  or school.If things were normal I would have done more for project. I'm that kinda person that I have to be around classmates or people to get ideas and work done. So isolation  is been difficult for me in my learning journey .